adjust all models types and associations

+++ packages/pomodoro-api/models/type.js
@@ -1,7 +1,7 @@
 module.exports = (sequelize, DataTypes) => {
   const Type = sequelize.define('Type', {
     name: DataTypes.STRING,
-    duration: DataTypes.INTEGER,
+    duration: DataTypes.BIGINT,
   })
 
   Type.associate = (models) => {
